adpated jacobian computation to new code architecture
