+++ batchglm/train/tf/base_glm/__init__.py
@@ -1,3 +1,4 @@
 from .estimator_graph import GradientGraphGLM, NewtonGraphGLM, TrainerGraphGLM, EstimatorGraphGLM, FullDataModelGraphGLM
+from .jacobians import JacobiansGLM
 from .hessians import HessiansGLM
 from .model import ESTIMATOR_PARAMS, ProcessModelGLM, ModelVarsGLM, BasicModelGraphGLM
