+++ batchglm/train/tf/base_glm/jacobians.py
@@ -8,33 +8,37 @@ from .model import ModelVarsGLM
 logger = logging.getLogger(__name__)
 
 
-class HessiansGLM:
+class JacobiansGLM:
     """
-    Wrapper to compute the Hessian matrix for a GLM.
+    Wrapper to compute the Jacobian matrix for a GLM.
     """
 
-    hessian: tf.Tensor
-    neg_hessian: tf.Tensor
+    jac: tf.Tensor
+    neg_jac: tf.Tensor
 
     def __init__(
             self,
             batched_data: tf.data.Dataset,
             sample_indices: tf.Tensor,
+            batch_model,
             constraints_loc,
             constraints_scale,
             model_vars: ModelVarsGLM,
             noise_model: str,
             dtype,
-            mode="obs",
-            iterator=True,
-            hess_a=True,
-            hess_b=True,
+            mode="analytic",
+            iterator=False,
+            jac_a=True,
+            jac_b=True
     ):
-        """ Return computational graph for hessian based on mode choice.
+        """ Return computational graph for jacobian based on mode choice.
 
         :param batched_data:
             Dataset iterator over mini-batches of data (used for training) or tf.Tensors of mini-batch.
-        :param sample_indices: Indices of samples to be used.
+        :param sample_indices:
+            Indices of samples to be used.
+        :param batch_model: BasicModelGraph instance
+            Allows evaluation of jacobian via tf.gradients as it contains model graph.
         :param constraints_loc: np.ndarray (constraints on mean model x mean model parameters)
             Constraints for location model.
             Array with constraints in rows and model parameters in columns.
@@ -55,50 +59,36 @@ class HessiansGLM:
             parameter is indicated by a -1 in this array, the independent parameters
             of that constraint (which may be dependent at an earlier constraint)
             are indicated by a 1.
-        :param model_vars: TODO
-        :param noise_model: str {"nb"}
-            Noise model identifier.
         :param dtype: Precision used in tensorflow.
         :param mode: str
             Mode by with which hessian is to be evaluated,
-            for analytic solutions of the hessian one can either chose by
-            "feature" or by "obs" (observation). Note that sparse
-            observation matrices X are often csr, ie. slicing is
-            faster by row/observation, so that hessian evaluation
-            by observation is much faster. "tf" allows for
-            evaluation of the hessian via the tf.hessian function,
-            which is done by feature for implementation reasons.
+            "analytic" uses a closed form solution of the jacobian,
+            "tf" allows for evaluation of the jacobian via the tf.gradients function.
         :param iterator: bool
-            Whether batched_data is an iterator or a tensor (such as single yield of an iterator).
-        :param hess_a: bool
-            Wether to compute Hessian block for a parameters. If both hess_a and hess_b are true,
-            the entire hessian with the off-diagonal a-b block is computed in self.hessian.
-        :param hess_b: bool
-            Wether to compute Hessian block for b parameters. If both hess_a and hess_b are true,
-            the entire hessian with the off-diagonal a-b block is computed in self.hessian.
+            Whether an iterator or a tensor (single yield of an iterator) is given
+            in.
+        :param jac_a: bool
+            Wether to compute Jacobian for a parameters. If both jac_a and jac_b are true,
+            the entire jacobian is computed in self.jac.
+        :param jac_b: bool
+            Wether to compute Jacobian for b parameters. If both jac_a and jac_b are true,
+            the entire jacobian is computed in self.jac.
         """
         if constraints_loc is not None and mode != "tf":
-            raise ValueError("closed form hessian does not work if constraints_loc is not None")
+            raise ValueError("closed form jacobians do not work if constraints_loc is not None")
         if constraints_scale is not None and mode != "tf":
-            raise ValueError("closed form hessian does not work if constraints_scale is not None")
+            raise ValueError("closed form jacobians do not work if constraints_scale is not None")
+
+        logger.debug("jacobian mode: %s" % mode)
+        logger.debug("compute jacobian for a model: %s" % str(jac_a))
+        logger.debug("compute jacobian for b model: %s" % str(jac_b))
 
         self.noise_model = noise_model
-        self._compute_hess_a = hess_a
-        self._compute_hess_b = hess_b
+        self._compute_jac_a = jac_a
+        self._compute_jac_b = jac_b
 
-        if mode == "obs_batched":
-            H = self.byobs(
-                batched_data=batched_data,
-                sample_indices=sample_indices,
-                constraints_loc=constraints_loc,
-                constraints_scale=constraints_scale,
-                model_vars=model_vars,
-                batched=True,
-                iterator=iterator,
-                dtype=dtype
-            )
-        elif mode == "feature":
-            H = self.byfeature(
+        if mode == "analytic":
+            J = self.analytic(
                 batched_data=batched_data,
                 sample_indices=sample_indices,
                 constraints_loc=constraints_loc,
@@ -108,9 +98,13 @@ class HessiansGLM:
                 dtype=dtype
             )
         elif mode == "tf":
-            H = self.tf_byfeature(
+            # Tensorflow computes the jacobian based on the objective,
+            # which is the negative log-likelihood. Accordingly, the jacobian
+            # is the negative jacobian computed here.
+            J = self.tf(
                 batched_data=batched_data,
                 sample_indices=sample_indices,
+                batch_model=batch_model,
                 constraints_loc=constraints_loc,
                 constraints_scale=constraints_scale,
                 model_vars=model_vars,
@@ -118,52 +112,41 @@ class HessiansGLM:
                 dtype=dtype
             )
         else:
-            raise ValueError("mode not recognized in hessian_nb_glm: " + mode)
+            raise ValueError("mode not recognized in Jacobian: " + mode)
 
         # Assign jacobian blocks.
         p_shape_a = model_vars.a_var.shape[0]  # This has to be _var to work with constraints.
-        if self._compute_hess_a and self._compute_hess_b:
-            H_a = H[:, :p_shape_a, :p_shape_a]
-            H_b = H[:, p_shape_a:, p_shape_a:]
-            negH = tf.negative(H)
-            negH_a = tf.negative(H_a)
-            negH_b = tf.negative(H_b)
-        elif self._compute_hess_a and not self._compute_hess_b:
-            H_a = H
-            H_b = None
-            negH = tf.negative(H)
-            negH_a = tf.negative(H_a)
-            negH_b = None
-        elif not self._compute_hess_a and self._compute_hess_b:
-            H_a = None
-            H_b = H
-            negH = tf.negative(H)
-            negH_a = None
-            negH_b = tf.negative(H_b)
+        if self._compute_jac_a and self._compute_jac_b:
+            J_a = J[:, :p_shape_a]
+            J_b = J[:, p_shape_a:]
+            negJ = tf.negative(J)
+            negJ_a = tf.negative(J_a)
+            negJ_b = tf.negative(J_b)
+        elif self._compute_jac_a and not self._compute_jac_b:
+            J_a = J
+            J_b = None
+            J = J
+            negJ = tf.negative(J)
+            negJ_a = tf.negative(J_a)
+            negJ_b = None
+        elif not self._compute_jac_a and self._compute_jac_b:
+            J_a = None
+            J_b = J
+            J = J
+            negJ = tf.negative(J)
+            negJ_a = None
+            negJ_b = tf.negative(J_b)
         else:
             raise ValueError("either require jac_a or jac_b")
 
-        self.hessian = H
-        self.hessian_aa = H_a
-        self.hessian_bb = H_b
-        self.neg_hessian = negH
-        self.neg_hessian_aa = negH_a
-        self.neg_hessian_bb = negH_b
+        self.jac = J
+        self.jac_a = J_a
+        self.jac_b = J_b
+        self.neg_jac = negJ
+        self.neg_jac_a = negJ_a
+        self.neg_jac_b = negJ_b
 
-    def byobs(
-            self,
-            batched_data,
-            sample_indices,
-            constraints_loc,
-            constraints_scale,
-            model_vars: ModelVarsGLM,
-            batched,
-            iterator,
-            dtype
-    ):
-        raise NotImplementedError()
-
-    def byfeature(
+    def analytic(
             self,
             batched_data,
             sample_indices,
@@ -175,10 +158,11 @@ class HessiansGLM:
     ):
         raise NotImplementedError()
 
-    def tf_byfeature(
+    def tf(
             self,
             batched_data,
             sample_indices,
+            batch_model,
             constraints_loc,
             constraints_scale,
             model_vars: ModelVarsGLM,
