+++ batchglm/train/tf/base_glm_all/estimator_graph.py
@@ -149,6 +149,7 @@ class FullDataModelGraph(FullDataModelGraphGLM):
                 constraints_scale=constraints_scale,
                 model_vars=model_vars,
                 mode=pkg_constants.JACOBIAN_MODE,
+                noise_model=noise_model,
                 iterator=True,
                 jac_a=True,
                 jac_b=True,
@@ -164,6 +165,7 @@ class FullDataModelGraph(FullDataModelGraphGLM):
                     constraints_scale=constraints_scale,
                     model_vars=model_vars,
                     mode=pkg_constants.JACOBIAN_MODE,
+                    noise_model=noise_model,
                     iterator=True,
                     jac_a=train_a,
                     jac_b=train_b,
@@ -362,6 +364,7 @@ class EstimatorGraphAll(EstimatorGraphGLM):
                     constraints_scale=constraints_scale,
                     model_vars=model_vars,
                     mode=pkg_constants.JACOBIAN_MODE,
+                    noise_model=noise_model,
                     iterator=False,
                     jac_a=train_loc,
                     jac_b=train_scale,
